refactor: update pages to use lib/cookies for auth
